[TensorExpr] Introduce ExternalCall nodes to TE IR.
ExternalCall nodes represent opaque calls to external functions to fill a
tensor (buffer) with values. It could be used to include nodes that are
otherwise not-representable as TE, or whose TE representation is currently too
slow.

To make an external function available in NNC as ExternalCall, one needs to
implement a "bridge" function that would take raw (void*) pointers to the data
along with the arrays containing dimension info. This function would then
internally call the desired external function and make sure the results of the
call are correctly placed in the provided raw data buffers.

ghstack-source-id: eb35102786c4907dc174aa8e18c934f7175fed4f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51475